+++ test/worker/worker2.js
@@ -1,5 +1,5 @@
 var counter = 100;
-function main(req, res, page) {
+local.worker.hostPage.handleRemoteWebRequest = function(req, res, page) {
 	if (req.path == '/' && req.method == 'GET') {
 		res.writeHead(200, 'ok', { 'content-type': 'text/plain' });
 		res.end(counter--);
@@ -12,5 +12,13 @@ function main(req, res, page) {
 		});
 		return;
 	}
+	if (req.path == '/' && req.method == 'BOUNCE') {
+		local.web.dispatch({ method: 'GET', url: 'httpl://0.page?foo='+local.worker.config.myname, query: { bar: 'buzz' } })
+			.always(function(res2) {
+				res.writeHead(200, 'ok', { 'content-type': 'text/plain' });
+				res.end(res2.body);
+			});
+		return;
+	}
 	res.writeHead(404, 'not found').end();
-}
+};